rename getSealCred
