+++ src/helpers/sealCred.ts
@@ -1,6 +1,6 @@
 import defaultProvider from 'helpers/defaultProvider'
-import getSealCred from 'helpers/getSealCred'
+import getSealCredERC721 from 'helpers/getSealCredERC721'
 import getSealCredEmail from 'helpers/getSealCredEmail'
 
-export const erc721Ledger = getSealCred(defaultProvider)
+export const erc721Ledger = getSealCredERC721(defaultProvider)
 export const workLedger = getSealCredEmail(defaultProvider)
